requirements.txt
paho-mqtt==2.0.0
pyserial==3.5
dsmr-parser==1.3.2
simplejson
requests